feat: logout, display user details
