+++ src/main/java/com/virtualclassroom/controller/AppController.java
@@ -1,15 +1,23 @@
 package com.virtualclassroom.controller;
 
 import com.virtualclassroom.model.User;
-import org.springframework.security.authentication.AnonymousAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
+import com.virtualclassroom.service.user.UserService;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class AppController {
+    private final UserService userService;
+
+    public AppController(UserService userService) {
+        this.userService = userService;
+    }
+
     @GetMapping("/")
     public String index() {
         return "index";
@@ -23,6 +31,15 @@ public class AppController {
     @GetMapping("/register")
     public String getRegisterPage(Model model) {
         model.addAttribute("user", new User());
+        return "login-register";
+    }
+
+    @PostMapping("/process_register")
+    public String processRegister(@NotNull User user) {
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+        String encodedPassword = passwordEncoder.encode(user.getUserPassword());
+        user.setUserPassword(encodedPassword);
+        userService.addUser(user);
         return "login-register";
     }
