+++ src/main/resources/templates/users.html
@@ -1,34 +1,40 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
+<html lang="en"
+      layout:decorate="_layout"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      xmlns:th="http://www.thymeleaf.org">
+<head th:replace="_layout :: head(type='tag')">
     <title>User Page</title>
 </head>
-<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+<link crossorigin="anonymous" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
+      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" rel="stylesheet">
 <body>
-<div class="container">
-    <div class="row">
-        <h1>List Users</h1>
+<div layout:fragment="main_body">
+    <br/>
+    <div class="container">
+        <div class="row">
+            <h1>List Users</h1>
+        </div>
+        <table class="table table-striped">
+            <thead>
+            <tr>
+                <th>ID</th>
+                <th>Name</th>
+                <th>Email</th>
+                <th>Roles</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr th:each="user: ${users}">
+                <td th:text="${user.id}"></td>
+                <td th:text="${user.userName}"></td>
+                <td th:text="${user.userEmail}"></td>
+                <td th:text="${user.roles}"></td>
+            </tr>
+            </tbody>
+        </table>
     </div>
-    <table class="table table-striped">
-       <thead>
-       <tr>
-           <th>ID</th>
-           <th>Name</th>
-           <th>Email</th>
-           <th>Roles</th>
-       </tr>
-       </thead>
-        <tbody>
-        <tr th:each = "user: ${users}">
-            <td th:text="${user.id}"></td>
-            <td th:text="${user.userName}"></td>
-            <td th:text="${user.userEmail}"></td>
-            <td th:text="${user.roles}"></td>
-        </tr>
-        </tbody>
-    </table>
+    <br/>
 </div>
 </body>
 </html>
